final commit

+++ src/Controller/UtilisateurController.php
@@ -11,7 +11,6 @@ use Symfony\Component\HttpFoundation\Request;
 use Doctrine\Persistence\ManagerRegistry;
 use App\Repository\UtilisateurRepository;
 use App\Form\UtilisateurType;
-use Container9VsbLlX\getMercuryseriesFlashy_FlashyNotifierService;
 use Doctrine\ORM\EntityManagerInterface;
 use MercurySeries\FlashyBundle\FlashyNotifier;
 use Symfony\Component\HttpFoundation\RedirectResponse;
@@ -21,32 +20,30 @@ use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 class UtilisateurController extends AbstractController
 {
 
-    private $session;
-
-    public function __construct(SessionInterface $session)
-    {
-        $this->session = $session;
-    }
-
     #[Route('/utilisateur', name: 'app_utilisateur')]
     public function index(): Response
     {
+
         return $this->render('utilisateur/index.html.twig', [
             'controller_name' => 'UtilisateurController',
         ]);
     }
 
     #[Route('/UsersListe', name: 'readUsers')]
-    public function listeUsers(UtilisateurRepository $repo): Response
+    public function listeUsers(UtilisateurRepository $repo, SessionInterface $session): Response
     {
-        $utilisateurs = $repo->findAll();
-        return $this->render('utilisateur/index.html.twig', [
-            'users' => $utilisateurs
-        ]);
+        if ($session->get('user')) {
+            $utilisateurs = $repo->findAll();
+            return $this->render('utilisateur/index.html.twig', [
+                'users' => $utilisateurs
+            ]);
+        } else {
+            return $this->render('notfound.html.twig');
+        }
     }
 
     #[Route('/addUser', name: 'signup')]
-    public function addUtilisateur(FlashyNotifier $flashy, ManagerRegistry $doctrine,  UserPasswordEncoderInterface $userPasswordEncoder, Request $req)
+    public function addUtilisateur(ManagerRegistry $doctrine,  UserPasswordEncoderInterface $userPasswordEncoder, Request $req, SessionInterface $session)
     {
         $user = new Utilisateur();
         $form = $this->createForm(UtilisateurType::class, $user);
@@ -65,7 +62,7 @@ class UtilisateurController extends AbstractController
             $em = $doctrine->getManager();
             $em->persist($user);
             $em->flush();
-            return $this->redirectToRoute('app_candidatures');
+            return $this->redirectToRoute('login');
         }
 
         return $this->render('utilisateur/signup.html.twig', [
@@ -73,43 +70,55 @@ class UtilisateurController extends AbstractController
         ]);
     }
 
+
     #[Route('/deleteUser/{id}', name: 'deleteUser')]
-    public function deleteUser($id, ManagerRegistry $doctrine, FlashyNotifier $flashy)
+    public function deleteUser($id, ManagerRegistry $doctrine, FlashyNotifier $flashy, SessionInterface $session)
     {
-        $utilisateur = $doctrine->getRepository(Utilisateur::class)->find($id);
-        $em = $doctrine->getManager();
-        $em->remove($utilisateur);
-        $em->flush();
-        $flashy->primaryDark('User deleted succesfully');
-        return $this->redirectToRoute('readUsers');
+        if ($session->get('user')) {
+            $utilisateur = $doctrine->getRepository(Utilisateur::class)->find($id);
+            $em = $doctrine->getManager();
+            $em->remove($utilisateur);
+            $em->flush();
+            $flashy->primaryDark('User deleted succesfully');
+            return $this->redirectToRoute('readUsers');
+        } else {
+            return $this->render('notfound.html.twig');
+        }
     }
 
     #[Route('/updateUser/{id}', name: 'updateUser')]
-    public function updateUser(Request $req, $id, ManagerRegistry $doctrine, FlashyNotifier $flashy)
+    public function updateUser(Request $req, $id, ManagerRegistry $doctrine, FlashyNotifier $flashy, SessionInterface $session)
     {
-        $utilisateur = $doctrine->getRepository(Utilisateur::class)->find($id);
-        $form = $this->createForm(UtilisateurType::class, $utilisateur);
-        $form->handleRequest($req);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $doctrine->getManager();
-            $em->flush();
-            $flashy->primaryDark('User updated succesfully');
-            return $this->redirectToRoute('readUsers');
+        if ($session->get('user')) {
+            $utilisateur = $doctrine->getRepository(Utilisateur::class)->find($id);
+            $form = $this->createForm(UtilisateurType::class, $utilisateur);
+            $form->handleRequest($req);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $doctrine->getManager();
+                $em->flush();
+                $flashy->primaryDark('User updated succesfully');
+                return $this->redirectToRoute('readUsers');
+            }
+
+            return $this->render('utilisateur/updateUtilisateur.html.twig', [
+                'form' => $form->createView()
+            ]);
+        } else {
+            return $this->render('notfound.html.twig');
         }
-
-        return $this->render('utilisateur/updateUtilisateur.html.twig', [
-            'form' => $form->createView()
-        ]);
     }
+
+
     #[Route('/connexion', name: 'login')]
-    public function login(FlashyNotifier $flashy, UtilisateurRepository $userRepository, UserPasswordEncoderInterface $userPasswordEncoder,  Request $req, EntityManagerInterface $entityManager, SessionInterface $session): Response
+    public function login(UtilisateurRepository $userRepository, UserPasswordEncoderInterface $userPasswordEncoder,  Request $req, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
         $error = '';
         $form = $this->createForm(LoginFormType::class);
         $form->handleRequest($req);
         if ($form->isSubmitted() && $form->isValid()) {
-            // $data = $form->getData();
+
             // Récupérer l'utilisateur correspondant à l'e-mail entré
             $email = $form->get('email')->getData();
             $plainPassword = $form->get('mdp')->getData();
@@ -118,16 +127,21 @@ class UtilisateurController extends AbstractController
             // Vérifier si le mot de passe entré correspond à celui stocké dans la base de données
             if ($user != null && $userPasswordEncoder->isPasswordValid($user, $plainPassword)) {
                 $session->set('user', $user);
-                dump('Authentification réussie');
-                $flashy->primaryDark('Log in succesfully');
+                /* dump($this->session->get('user', $user));
+                die();*/
+
                 // Authentification réussie, redirection
-                if ($user->getIdrole()->getDescription() == 'Administrateur') {
+                if ($session->get('user')->getIdrole()->getDescription() == 'Administrateur') {
+
                     return $this->redirectToRoute('readUsers');
-                } else
+                } else {
+
                     return $this->redirectToRoute('app_candidatures');
+                }
             } else {
+
+
                 $error = 'Adresse e-mail ou mot de passe incorrect';
-                $flashy->error('erreur lors de l\'authentification');
                 dump($error);
             }
         }
@@ -137,13 +151,12 @@ class UtilisateurController extends AbstractController
     }
 
     #[Route('/logout', name: 'logout')]
-    public function logout(FlashyNotifier $flashy, SessionInterface $session, Request $req, UtilisateurRepository $userRepository)
+    public function logout(SessionInterface $session)
     {
         $session->remove('user');
 
         // Rediriger l'utilisateur vers la page d'accueil après la déconnexion
         $response = new RedirectResponse('/connexion');
         return $response;
-        $flashy->primaryDark('Déconnexion');
     }
 }
